+++ test_runner/src/main/kotlin/ftl/reports/MatrixResultsReport.kt
@@ -4,6 +4,7 @@ import ftl.args.IArgs
 import ftl.config.FtlConstants.indent
 import ftl.gc.GcStorage
 import ftl.json.MatrixMap
+import ftl.json.SavedMatrix
 import ftl.reports.util.IReport
 import ftl.reports.xml.model.JUnitTestResult
 import ftl.util.asPrintableTable
@@ -48,13 +49,23 @@ object MatrixResultsReport : IReport {
             }
 
             if (matrices.map.isNotEmpty()) {
-                writer.println("More details are available at [${matrices.map.values.first().webLinkWithoutExecutionDetails}]")
-                writer.println(matrices.map.values.toList().asPrintableTable())
+                val savedMatrices = matrices.map.values
+                writer.println(savedMatrices.toList().asPrintableTable())
+                savedMatrices.printMatricesLinks(writer)
             }
 
             return writer.toString()
         }
     }
+
+private fun Collection<SavedMatrix>.printMatricesLinks(writer: StringWriter) = this
+        .filter { it.failed() }
+        .takeIf { it.isNotEmpty() }
+        ?.run {
+            writer.println("More details are available at:")
+            forEach { writer.println(it.webLinkWithoutExecutionDetails) }
+            writer.println()
+        }
 
     override fun run(matrices: MatrixMap, result: JUnitTestResult?, printToStdout: Boolean, args: IArgs) {
         val output = generate(matrices)